feat : login activity mvvm migration
-> created ClientRepository, its implementation and unit tests.
-> removed Client related methods from UserAuthRepository, and its unit tests.
-> injected ClientRepositoryImp to LoginViewModel through Hilt, and added its entry to repository module of DI
