+++ app/src/main/java/org/mifos/mobile/repositories/UserAuthRepository.kt
@@ -2,9 +2,7 @@ package org.mifos.mobile.repositories
 
 import io.reactivex.Observable
 import okhttp3.ResponseBody
-import org.mifos.mobile.models.Page
 import org.mifos.mobile.models.User
-import org.mifos.mobile.models.client.Client
 
 interface UserAuthRepository {
 
@@ -20,15 +18,4 @@ interface UserAuthRepository {
     ): Observable<ResponseBody?>?
 
     fun login(username: String, password: String): Observable<User?>?
-
-    fun loadClient() : Observable<Page<Client?>?>?
-
-    fun saveAuthenticationTokenForSession(user: User)
-
-    fun reInitializeService()
-
-    fun setClientId(clientId: Long?)
-
-    fun clearPrefHelper()
-
 }
