+++ app/src/test/java/org/mifos/mobile/repositories/UserAuthRepositoryImpTest.kt
@@ -8,9 +8,7 @@ import org.junit.Test
 import org.junit.runner.RunWith
 import org.mifos.mobile.FakeRemoteDataSource
 import org.mifos.mobile.api.DataManager
-import org.mifos.mobile.models.Page
 import org.mifos.mobile.models.User
-import org.mifos.mobile.models.client.Client
 import org.mifos.mobile.models.payload.LoginPayload
 import org.mifos.mobile.models.register.RegisterPayload
 import org.mockito.Mock
@@ -26,14 +24,13 @@ class UserAuthRepositoryImpTest {
 
     private lateinit var userAuthRepositoryImp: UserAuthRepository
     private lateinit var mockUser : User
-    private var mockClientPage : Page<Client?>? = null
+
 
     @Before
     fun setUp() {
         MockitoAnnotations.openMocks(this)
         userAuthRepositoryImp = UserAuthRepositoryImp(dataManager)
         mockUser = FakeRemoteDataSource.user
-        mockClientPage = FakeRemoteDataSource.clients
     }
 
     @Test
@@ -130,32 +127,6 @@ class UserAuthRepositoryImpTest {
         val result = userAuthRepositoryImp.login("username", "password")
 
         Mockito.verify(dataManager).login(mockLoginPayload)
-        Assert.assertEquals(result, errorResponse)
-    }
-
-    @Test
-    fun testLoadClient_SuccessResponseReceivedFromDataManager_ReturnsClientPageSuccessfully() {
-        val successResponse : Observable<Page<Client?>?> = Observable.just(mockClientPage)
-        Mockito.`when`(
-            dataManager.clients
-        ).thenReturn(successResponse)
-
-        val result = userAuthRepositoryImp.loadClient()
-
-        Mockito.verify(dataManager).clients
-        Assert.assertEquals(result, successResponse)
-    }
-
-    @Test
-    fun testLoadClient_ErrorResponseReceivedFromDataManager_ReturnsError() {
-        val errorResponse : Observable<Page<Client?>?> = Observable.error(Throwable("Load Client Failed"))
-        Mockito.`when`(
-            dataManager.clients
-        ).thenReturn(errorResponse)
-
-        val result = userAuthRepositoryImp.loadClient()
-
-        Mockito.verify(dataManager).clients
         Assert.assertEquals(result, errorResponse)
     }
 }
